Add files via upload
A triangle can be determined to be a right triangle if the side lengths are known. If the lengths satisfy the Pythagorean Theorem (a2+b2=c2) then it is a right triangle. Write a C program that gets the sides of triangle, a,b,c and check whether the triangle is right-angled or not.